requirements.txt
aiogram==3.6.0
asyncpg==0.29.0
python-dotenv==1.0.1